+++ src/mui/FolderRequestWindow.c
@@ -156,7 +156,7 @@ OVERLOAD(OM_NEW)
       End,
     End,
 
-    TAG_MORE, (ULONG)inittags(msg))) != NULL)
+    TAG_MORE, inittags(msg))) != NULL)
   {
     GETDATA;
