+++ src/mui/StringRequestWindow.c
@@ -31,7 +31,6 @@
 #include "StringRequestWindow_cl.h"
 
 #include "YAM_stringsizes.h"
-
 #include "Locale.h"
 #include "MUIObjects.h"
 
@@ -151,7 +150,7 @@ OVERLOAD(OM_NEW)
       End,
     End,
 
-    TAG_MORE, (ULONG)inittags(msg))) != NULL)
+    TAG_MORE, inittags(msg))) != NULL)
   {
     GETDATA;
